+++ Grapengine/controllers/ge_app_controller.hpp
@@ -18,7 +18,7 @@ namespace GE::Ctrl
 
     static void AllowImGuiEvents(bool value);
 
-    static f32 GetFPS();
+    static f64 GetFPS();
 
   private:
     App() = default;
